Converted camera to f32

+++ src/camera.rs
@@ -3,15 +3,15 @@ use na::{Isometry3, Perspective3, Point3, Vector3};
 
 pub struct Camera {
     pub stale: bool,
-    pub cache: Matrix4<f64>,
-    pub pos: Point3<f64>,
-    pub forward: Vector3<f64>,
-    pub up: Vector3<f64>,
-    pub persp: Perspective3<f64>,
+    pub cache: Matrix4<f32>,
+    pub pos: Point3<f32>,
+    pub forward: Vector3<f32>,
+    pub up: Vector3<f32>,
+    pub persp: Perspective3<f32>,
 }
 
 impl Camera {
-    pub fn new(pos: Point3<f64>, rot: Vector3<f64>, persp: Perspective3<f64>) -> Camera {
+    pub fn new(pos: Point3<f32>, rot: Vector3<f32>, persp: Perspective3<f32>) -> Camera {
         Camera {
             stale: true,
             cache: Matrix4::identity(),
@@ -22,7 +22,7 @@ impl Camera {
         }
     }
 
-    pub fn fresh_mat(&mut self) -> &Matrix4<f64> {
+    pub fn fresh_mat(&mut self) -> &Matrix4<f32> {
         if self.stale {
             self.cache = self.persp.as_matrix()
                 * Isometry3::face_towards(&self.pos, &(self.pos + self.forward), &self.up)
@@ -32,17 +32,17 @@ impl Camera {
         &self.cache
     }
 
-    pub fn right(&self) -> Vector3<f64> {
-        let ymp = 0.0f64; //self.rot.y - std::f64::consts::PI / 2.0;
+    pub fn right(&self) -> Vector3<f32> {
+        let ymp = 0.0f32; //self.rot.y - std::f32::consts::PI / 2.0;
         Vector3::new(ymp.sin(), 0.0, ymp.cos())
     }
 
-    pub fn resize(&mut self, size: [f64; 2]) {
+    pub fn resize(&mut self, size: [f32; 2]) {
         self.persp.set_aspect(size[0] / size[1]);
         self.stale = true;
     }
 
-    pub fn mov(&mut self, x: f64, y: f64, z: f64) {
+    pub fn mov(&mut self, x: f32, y: f32, z: f32) {
         self.pos += self.forward * z;
         self.pos += self.up * y;
         self.pos += self.right() * x;
